+++ lockfile_generator/src/yarn.rs
@@ -1,9 +1,10 @@
 //! JavaScript yarn ecosystem.
 
+use std::fs;
 use std::path::Path;
 use std::process::Command;
 
-use crate::Generator;
+use crate::{Error, Generator, Result};
 
 pub struct Yarn;
 
@@ -20,5 +21,31 @@ impl Generator for Yarn {
 
     fn tool(&self) -> &'static str {
         "Yarn"
+    }
+
+    fn check_prerequisites(&self, manifest_path: &Path) -> Result<()> {
+        let yarn_version = yarn_version(manifest_path)?;
+        if yarn_version.starts_with("1.") {
+            let version = yarn_version.trim().into();
+            return Err(Error::UnsupportedCommandVersion("yarn", "2.0.0+", version));
+        }
+
+        Ok(())
+    }
+}
+
+/// Get the yarn version of the project.
+fn yarn_version(manifest_path: &Path) -> Result<String> {
+    let canonicalized = fs::canonicalize(manifest_path)?;
+    let project_path = canonicalized
+        .parent()
+        .ok_or_else(|| Error::InvalidManifest(manifest_path.to_path_buf()))?;
+
+    let output = Command::new("yarn").arg("--version").current_dir(project_path).output()?;
+    if output.status.success() {
+        Ok(String::from_utf8_lossy(&output.stdout).into())
+    } else {
+        let stderr = String::from_utf8_lossy(&output.stderr);
+        Err(Error::NonZeroExit(output.status.code(), stderr.into()))
     }
 }