+++ CHANGELOG.md
@@ -14,6 +14,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Correctly handle line continuations and --hash in `requirements.txt` parser
 
+### Removed
+- Lockfile generation for yarn v1
+
 ## [5.3.0] - 2023-06-15
 
 ### Added